requirements.txt
pandas
numpy
matplotlib
python-binance
scikit-learn
ta
flask